// web-new-theme/src/app/main/admin/user-management/users/models/user.ts
export class User {
    userName: string;
    userId: number;
    firstName: string;
    lastName: string;
    status: string;
    userRole: string;
    fullName: string;
    employeeId: number;
    roleIds: any[];
    password: string;

    oragnizationId: number;

    constructor() {
        this.oragnizationId = 1;
    }

}